+++ src/registry.ts
@@ -1,33 +1,14 @@
 // src/registry.ts
-import { Server } from "@modelcontextprotocol/sdk/server/index.js"
 import type { StdioServerParameters } from "@modelcontextprotocol/sdk/client/stdio.js"
 import { REGISTRY_URL } from "./config.js"
 import { StdioClientTransport } from "@modelcontextprotocol/sdk/client/stdio.js"
+import type { RegistryPackage, RegistryVariables } from "./types.js"
 // import { SSEClientTransport } from "@modelcontextprotocol/sdk/client/sse.js"
-
-// Basic types for our registry
-export interface RegistryPackage {
-  id: string
-  name: string
-  description: string
-  vendor: string
-  sourceUrl: string
-  license: string
-  homepage: string
-  connections: Array<{
-    configSchema: Record<string, any>
-    stdio: {
-      command: string
-      args: string[]
-      env: Record<string, any>
-    }
-  }>
-}
 
 // Helper to create StdioClientTransport config
 export function createStdioConfig(
   pkg: RegistryPackage, 
-  variables: Record<string, string>
+  variables: RegistryVariables
 ): StdioServerParameters {
   if (pkg.connections.length === 0) {
     throw new Error(`No connections defined for package: ${pkg.id}`)
@@ -37,13 +18,15 @@ export function createStdioConfig(
   const connection = pkg.connections[0]
   
   const env: Record<string, string> = {}
-  for (const [key, template] of Object.entries(connection.stdio.env)) {
-    env[key] = template.toString().replace(/\${([^}]+)}/g, (_: string, varName: string) => {
-      if (!(varName in variables)) {
-        throw new Error(`Missing required variable: ${varName}`)
-      }
-      return variables[varName]
-    })
+  if (connection.stdio.env) {
+    for (const [key, template] of Object.entries(connection.stdio.env)) {
+      env[key] = template.toString().replace(/\${([^}]+)}/g, (_: string, varName: string) => {
+        if (!(varName in variables)) {
+          throw new Error(`Missing required variable: ${varName}`)
+        }
+        return variables[varName]
+      })
+    }
   }
 
   return {
@@ -70,7 +53,7 @@ export async function fetchRegistryEntry(
 
 export async function createTransport(
   id: string,
-  variables: Record<string, string>
+  variables: RegistryVariables
 ) {
   const pkg = await fetchRegistryEntry(id)
   if (!pkg) {
